// src/types.ts
export interface DifferenceEqual {
    /**
     * path not including file/directory name; can be relative or absolute depending on call to compare().
     */
    path1: string;

    /**
     * path not including file/directory name; can be relative or absolute depending on call to compare().
     */
    path2: string;

    /**
     * path relative to root.
     */
    relativePath: string;

    /**
     * file/directory name.
     */
    name1: string;

    /**
     * file/directory name.
     */
    name2: string;

    /**
     * one of equal, left, right, distinct.
     */
    state: 'equal';

    /**
     * one of missing, file, directory.
     */
    type1: 'file' | 'directory';

    /**
     * one of missing, file, directory.
     */
    type2: 'file' | 'directory';

    /**
     * file size.
     */
    size1: number;

    /**
     * file size.
     */
    size2: number;

    /**
     * modification date (stat.mtime).
     */
    date1: number;

    /**
     * modification date (stat.mtime).
     */
    date2: number;

    /**
     * depth.
     */
    level: number;
}

export interface DifferenceDistinct {
    path1: string;
    path2: string;
    relativePath: string;
    name1: string;
    name2: string;
    state: 'distinct';
    type1: 'file' | 'directory';
    type2: 'file' | 'directory';
    size1: number;
    size2: number;
    date1: number;
    date2: number;
    level: number;
}

export interface DifferenceLeft {
    path1: string;
    path2: undefined;
    relativePath: string;
    name1: string;
    name2: undefined;
    state: 'left';
    type1: 'file' | 'directory';
    type2: 'missing';
    size1: number;
    size2: undefined;
    date1: number;
    date2: undefined;
    level: number;
}

export interface DifferenceRight {
    path1: undefined;
    path2: string;
    relativePath: string;
    name1: undefined;
    name2: string;
    state: 'right';
    type1: 'missing';
    type2: 'file' | 'directory';
    size1: undefined;
    size2: number;
    date1: undefined;
    date2: number;
    level: number;
}